refactor : Separation of converter class from Restaurant
