+++ src/main/java/com/onsuyum/common/util/StringListConverter.java
@@ -1,0 +1,20 @@
+package com.onsuyum.common.util;
+
+import javax.persistence.AttributeConverter;
+import javax.persistence.Converter;
+import java.util.Arrays;
+import java.util.List;
+
+@Converter
+public class StringListConverter implements AttributeConverter<List<String>, String> {
+    private static final String SPLIT_CHAR = ";!";
+    @Override
+    public String convertToDatabaseColumn(List<String> attribute) {
+        return String.join(SPLIT_CHAR, attribute);
+    }
+
+    @Override
+    public List<String> convertToEntityAttribute(String dbData) {
+        return Arrays.asList(dbData.split(SPLIT_CHAR));
+    }
+}
